Extract repeated definitions of MockTracerListener to a single .h file
These were identical across the tests of the four `PerfEventVisitors`.
