+++ src/LinuxTracing/GpuTracepointVisitorTest.cpp
@@ -14,35 +14,15 @@
 
 #include "GpuTracepointVisitor.h"
 #include "KernelTracepoints.h"
+#include "MockTracerListener.h"
 #include "OrbitBase/Logging.h"
 #include "PerfEvent.h"
 #include "PerfEventRecords.h"
-#include "TracingInterface/TracerListener.h"
 #include "capture.pb.h"
 
 namespace orbit_linux_tracing {
 
 namespace {
-
-class MockTracerListener : public orbit_tracing_interface::TracerListener {
- public:
-  MOCK_METHOD(void, OnSchedulingSlice, (orbit_grpc_protos::SchedulingSlice), (override));
-  MOCK_METHOD(void, OnCallstackSample, (orbit_grpc_protos::FullCallstackSample), (override));
-  MOCK_METHOD(void, OnFunctionCall, (orbit_grpc_protos::FunctionCall), (override));
-  MOCK_METHOD(void, OnGpuJob, (orbit_grpc_protos::FullGpuJob full_gpu_job), (override));
-  MOCK_METHOD(void, OnThreadName, (orbit_grpc_protos::ThreadName), (override));
-  MOCK_METHOD(void, OnThreadNamesSnapshot, (orbit_grpc_protos::ThreadNamesSnapshot), (override));
-  MOCK_METHOD(void, OnThreadStateSlice, (orbit_grpc_protos::ThreadStateSlice), (override));
-  MOCK_METHOD(void, OnAddressInfo, (orbit_grpc_protos::FullAddressInfo), (override));
-  MOCK_METHOD(void, OnTracepointEvent, (orbit_grpc_protos::FullTracepointEvent), (override));
-  MOCK_METHOD(void, OnModuleUpdate, (orbit_grpc_protos::ModuleUpdateEvent), (override));
-  MOCK_METHOD(void, OnModulesSnapshot, (orbit_grpc_protos::ModulesSnapshot), (override));
-  MOCK_METHOD(void, OnErrorsWithPerfEventOpenEvent,
-              (orbit_grpc_protos::ErrorsWithPerfEventOpenEvent), (override));
-  MOCK_METHOD(void, OnLostPerfRecordsEvent, (orbit_grpc_protos::LostPerfRecordsEvent), (override));
-  MOCK_METHOD(void, OnOutOfOrderEventsDiscardedEvent,
-              (orbit_grpc_protos::OutOfOrderEventsDiscardedEvent), (override));
-};
 
 class GpuTracepointVisitorTest : public ::testing::Test {
  protected:
